[opentitanlib] Move USB backend implementation based on rusb
(cherry picked from commit 4f4cd775672a414a523e9690c6c96c7402953e67)
