+++ sw/host/opentitanlib/BUILD
@@ -164,6 +164,7 @@ rust_library(
         "src/tpm/status.rs",
         "src/transport/common/fpga.rs",
         "src/transport/common/mod.rs",
+        "src/transport/common/usb.rs",
         "src/transport/errors.rs",
         "src/transport/ioexpander/mod.rs",
         "src/transport/ioexpander/sx1503.rs",
@@ -188,7 +189,6 @@ rust_library(
         "src/util/serializable_error.rs",
         "src/util/status.rs",
         "src/util/testing.rs",
-        "src/util/usb.rs",
         "src/util/usr_access.rs",
         "src/util/vcd.rs",
         "src/util/vmem/mod.rs",
