+++ sw/host/opentitanlib/src/transport/common/mod.rs
@@ -3,3 +3,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod fpga;
+pub mod usb;
