+++ sw/host/opentitanlib/src/transport/common/usb.rs
