+++ sw/host/opentitanlib/src/util/mod.rs
@@ -19,7 +19,6 @@ pub mod serde;
 pub mod serializable_error;
 pub mod status;
 pub mod testing;
-pub mod usb;
 pub mod usr_access;
 pub mod vcd;
 pub mod vmem;
