+++ sw/host/tests/chip/usb/usbdev_suspend.rs
@@ -22,8 +22,8 @@ use opentitanlib::app::TransportWrapper;
 use opentitanlib::execute_test;
 use opentitanlib::io::uart::Uart;
 use opentitanlib::test_utils::init::InitializeTest;
+use opentitanlib::transport::common::usb::{UsbHub, UsbHubOp};
 use opentitanlib::uart::console::UartConsole;
-use opentitanlib::util::usb::{UsbHub, UsbHubOp};
 
 use usb::{UsbOpts, get_device_by_port_numbers};
